+++ translations/el_GR.ts
@@ -890,6 +890,49 @@ Press any key to exit</source>
     </message>
 </context>
 <context>
+    <name>FakeVimCommands</name>
+    <message>
+        <source>`new` requires no argument or one of &apos;cpp&apos;, &apos;java&apos; and &apos;python&apos;, got [%1]</source>
+        <translation>η `new` δεν δέχεται όρισμα ή πρέπει να είναι ένα από &apos;cpp&apos;, &apos;java&apos; και &apos;python&apos;, ελήφθη [%1]</translation>
+    </message>
+    <message>
+        <source>[%1] is not C++, Python or Java source file</source>
+        <translation>το [%1] δεν είναι αρχείο πηγαίου C++, Python ή Java</translation>
+    </message>
+    <message>
+        <source>[%1] does not exist. To open a tab with a non-existing file, use `open!` instead</source>
+        <translation>το [%1] δεν υπάρχει. Για να ανοίξετε καρτέλα με ανύπαρκτο αρχείο, χρησιμοποιήστε `open!`</translation>
+    </message>
+    <message>
+        <source>[%1] is not a number</source>
+        <translation>το [%1] δεν είναι αριθμός</translation>
+    </message>
+    <message>
+        <source>%1 is out of range [1, %2]</source>
+        <translation>το %1 είναι εκτός εύρους [1, %2]</translation>
+    </message>
+    <message>
+        <source>N/A</source>
+        <translation>Δ/Υ</translation>
+    </message>
+    <message>
+        <source>[%1] is not a valid view mode. It should be one of &apos;split&apos; and &apos;edit&apos;</source>
+        <translation>το [%1] δεν είναι έγκυρη λειτουργία προβολής. Πρέπει να είναι &apos;split&apos; ή &apos;edit&apos;</translation>
+    </message>
+    <message>
+        <source>%1 is not a valid language name. It should be one of &apos;cpp&apos;, &apos;java&apos; and &apos;python&apos;</source>
+        <translation>το %1 δεν είναι έγκυρο όνομα γλώσσας. Πρέπει να είναι &apos;cpp&apos;, &apos;java&apos; ή &apos;python&apos;</translation>
+    </message>
+    <message>
+        <source>No active tab to clear messages</source>
+        <translation>Δεν υπάρχει ενεργή καρτέλα για εκκαθάριση μηνυμάτων</translation>
+    </message>
+    <message>
+        <source>No active tab to change language</source>
+        <translation type="unfinished"></translation>
+    </message>
+</context>
+<context>
     <name>FindReplaceDialog</name>
     <message>
         <source>Find/Replace</source>
@@ -2792,6 +2835,24 @@ This may reduce distractions caused by stopwatch updates.</source>
     <message>
         <source>Path to the template file</source>
         <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Enable Vim Emulation</source>
+        <translation>Ενεργοποίηση εξομοίωσης Vim</translation>
+    </message>
+    <message>
+        <source>Enable vim emulation in Code Editor</source>
+        <translation>Ενεργοποίηση εξομοίωσης Vim στον Επεξεργαστή Κώδικα</translation>
+    </message>
+    <message>
+        <source>Vim Configuration</source>
+        <translation>Ρυθμίσεις Vim</translation>
+    </message>
+    <message>
+        <source>The contents of Vim RC. It is loaded everytime vim emulation starts. 
+Not all vim commands are supported, please check https://github.com/cpeditor/FakeVim for list of supported commands</source>
+        <translation>Το περιεχόμενο του Vim RC. Φορτώνεται κάθε φορά που ξεκινά η εξομοίωση Vim.
+Δεν υποστηρίζονται όλες οι εντολές Vim, ελέγξτε https://github.com/cpeditor/FakeVim για τη λίστα υποστηριζόμενων εντολών</translation>
     </message>
 </context>
 <context>